Adding a file watcher for changes in the filesystem.
The file watcher is used via library (https://github.com/gmethvin/directory-watcher)

Fixed a few tests to use changes in the filesystem for deleting and updating files from
the index.

The code is quite messy at the moment. It requires a bit of refactoring. And probably
re-implementation of the watching loop to be more coroutine friendly.