Fixed bug with certain terms not appearing in search results.
The bug was introduced in 7019107891bd74a31b231856fb5dbeaafc91cd13 when trying
to allow searching while indexing.

The initial batch indexing happens in parallel. Before that change, the indexer
waited until all the batchs were processed and then added the shards to the
index.

With the change to allow searching while indexing, the shards where added by
each worker immediately after the shard is processed. But because there are
multiple jobs creating shards simultaneously, it is possible that the shards
end up added in different document ID order.

Having the shards ordered by documentID is essential for some QueryResolvers to
work correctly. For example, the `TrigramSubstringQueryResolver` was relying on
this to be able to merge multiple search results in linear time. When the
shards got order got messed up, the merging function was producing incorrect
results. Resulting in missng documents from the results.

This fixes the issue by ensuring that every time that a shard is added, the list
of shards is ordered by the first document ID appearing in the shard.